+++ src/test/java/test/HomeFaqTest.java
@@ -13,7 +13,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import pageObjects.HomeFaqPage;
+import page.objects.HomeFaqPage;
 
 import java.time.Duration;
