+++ src/test/java/test/OrderFormTest.java
@@ -13,7 +13,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
-import pageObjects.OrderForm;
+import page.objects.OrderForm;
 
 import java.time.Duration;
